Added "cursor: pointer" CSS to blocks and tabs in the workspace to indicate that they can be interacted with.
Text fields in the "palette" in the workspace are also given this class, since their contents can't be manipulated while in the palette (they are merely drag handles like the rest of the block they're in).

Also gave block labels "user-select: none" CSS to ensure that the label text doesn't get selected if the user misses a block when click-dragging.